[issue460] use double for search max_time

+++ src/search/search_engine.cc
@@ -14,7 +14,7 @@ SearchEngine::SearchEngine(const Options &opts)
       solution_found(false),
       search_space(OperatorCost(opts.get_enum("cost_type"))),
       cost_type(OperatorCost(opts.get_enum("cost_type"))),
-      max_time(opts.get<int>("max_time")) {
+      max_time(opts.get<double>("max_time")) {
     if (opts.get<int>("bound") < 0) {
         cerr << "error: negative cost bound " << opts.get<int>("bound") << endl;
         exit_with(EXIT_INPUT_ERROR);
@@ -87,7 +87,7 @@ void SearchEngine::add_options_to_parser(OptionParser &parser) {
         "bound",
         "exclusive depth bound on g-values. Cutoffs are always performed according to "
         "the real cost, regardless of the cost_type parameter", "infinity");
-    parser.add_option<int>(
+    parser.add_option<double>(
         "max_time",
         "maximum time in seconds the search is allowed to run for. The "
         "timeout is only checked after each complete search step "
